Fix compilation warning on OSX.
Get two warning when compiling for UE4 Editor on Mac. I simply rename the variable that was unhappy with the shadowing  of variable. But for tell the true I think it is because it is a old version of clang.

clip_reader.h:371:48: error: declaration shadows a local variable [-Werror,-Wshadow]
clip_writer.h:50:46: error: declaration shadows a local variable [-Werror,-Wshadow]