perf(decompression): Optimize the decompression seek function
The old function performed a linear search to find the segments
we need to sample a time T. This can be slow for long clips.
Instead we now make an approximate guess where the first
segment we need lives since the segment sizes are near uniform.
This allows us to only check neighboring segments in O(1).

Fixes #185