perf(decompression): convert flag arrays into a bit field
Bool flag arrays ended up on the stack with read/writes happening there as well.
On x64, instructions exist to work directly from memory and the asm generated
was short and concise but that wasn't the case on ARM. Instead, we now use a
bit field which nicely fits into a single register. This leads to shorter and faster
assembly.

Fixes #193