perf(decompression): reduce stack usage and unroll loops explicitly
With MSVC, having an array of vector4 on the stack causes stack
cookies to be introduced which slows things down. The compilers
also fail to keep values in registers when they otherwise could due
to unrolling that already happens automatically. To resolve this,
we use the stack via the sampling context when looping and
we switch to registers when we can unroll safely.

Fixes #194
Fixes #195