fix: fix utility functions that deal with num sample and duration conversion
Previously, when converting a duration to a number of samples, a duration of 0.0 seconds would yield a
single sample when it should be 0. An infinite duration now returns a single sample (a static pose).
Similarly, when calculating a duration from the number of samples we properly handle 0 and 1 to yield
a duration of 0.0 and infinity respectively.