feat(compression): add a track database to cache bit rate queries
We query the same bit rates over and over as we iterate to optimize. We now cache the result aggressively for
the least recently used 4 bit rates for each track rot/trans/scale. This avoids a massive amount of work.
Also split the error calculation function into two to avoid branching in the loop and for cleaner code.

Fixes #200