feat: WIP for clip metadata packing
Overall it yields a net loss. Even though we can significantly cut down the clip metadata cost,
animated samples are forced to retain more bits to compensate the accuracy loss.
Only implemented for rotation data, translation/scale require much complexity to handle
the fact that the range is much more dynamic. To retain enough accuracy there, we would
need something like K-means clusters. The added complexity and the higher memory footprint
make this bad idea for the average clip. It might be worth revisiting later for very short clips
with 10-15 frames or less. Perhaps this would be a win for single frame clips.