feat: retain constant rotation W components
Quantizing the constant rotation data yielded a net loss for the memory footprint.
The idea was to try the opposite and retain more precision by keeping the W
component we drop. This ended up yielding a very small net loss. It seems that
dropping the W component is the sweet spot.