feat(decompression): optimized the decompression to work in stages
General idea explained in #119. We now work with data that fits in one cache line at a time.
This allows us to manually prefetch the next cache line. It also allows us to unroll small loops
and process 4 samples at a time. This also makes it easier to use Structures of Arrays and
leverage the full SIMD width.