Fix uninitialized value in write_format_per_track_data
In the group_flush_action() lambda, the group_size var is overridden by 4 in the case of a rotation. In most cases that's fine because we'll memset the next group to 0. But if we have a single rotation group with less than 4 entries, then we rely on the uninitialized values from the stack.